# server/openapi.yaml
openapi: 3.1.0
...
info:
  title: Driftora food-parse
  version: '1.0.0'
  description: |
    Turns free text, photos and voice into structured nutrition and workouts.

    Stateless: nothing is persisted. Photos and audio clips live in memory for
    the duration of the request and are never written to disk (privacy §2).
    Request content is never logged — only aggregate counters.

    Energy numbers are estimates. Where the model reads a printed nutrition
    panel it is transcribing ground truth; everywhere else the client shows the
    result with «≈».
servers:
  - url: https://food.family-pie.ru
    description: production (single VPS behind Caddy)
  - url: http://127.0.0.1:8787
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /billing/register:
    post:
      tags: [billing]
      summary: Bind a store purchase to this install
      description: |
        Verifies the purchase with the store and, if it is real, raises this
        install's daily AI cap to the paid tier. `X-Install-Id` is required —
        it is what the entitlement is bound to.

        The client calls this after a purchase AND on every launch: re-registering
        is how a subscription follows the user onto a new phone (the install id
        changes, the store account's purchase does not), and how a refund ends
        access here — the service stores only a hash of the token and cannot
        re-ask the store on its own.

        `503 store_unreachable` does NOT mean the purchase is bad; retry later
        and do not tell the user their payment failed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [purchaseToken, productId]
              properties:
                purchaseToken: { type: string }
                productId:
                  type: string
                  description: |
                    A hint only. The stored tier comes from the product the STORE
                    reports for this token, never from this field.
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                type: object
                required: [active, expires_at, state]
                properties:
                  active: { type: boolean }
                  expires_at: { type: integer, description: Epoch ms; 0 when the store did not say. }
                  state: { type: string, enum: [active, grace, paused, expired, revoked] }
        '400': { description: Missing install id or malformed body }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { description: The store does not recognize this purchase }
        '503': { description: Billing is not configured here, or the store could not be reached }
...
  /billing/yookassa/webhook:
    post:
      tags: [billing]
      summary: ЮKassa payment notification
      description: |
        Called by ЮKassa, not by the app — so it is NOT token-authenticated.
        Authentication is the source-IP allowlist ЮKassa prescribes, and the
        handler re-reads the payment through the ЮKassa API rather than trusting
        the notification body (anyone able to spoof an allowed address could
        otherwise mint subscriptions by POSTing `payment.succeeded`).

        Status codes are a protocol: 200 stops redelivery, anything else makes
        ЮKassa retry for up to 24 hours. A payment we could not verify therefore
        answers 500 on purpose.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, example: notification }
                event: { type: string, example: payment.succeeded }
                object:
                  type: object
                  properties:
                    id: { type: string }
      responses:
        '200': { description: Processed, or understood and intentionally ignored }
        '400': { description: Notification without an object id }
        '403': { description: Not a ЮKassa source address }
        '500': { description: Could not verify the payment — ЮKassa should retry }
        '503': { description: Direct sales are not configured on this deployment }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `APP_TOKEN`, compared in constant time. When the variable is unset the
        check is a no-op and every route is public — acceptable only behind a
        network boundary.
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthorized, message: Missing or invalid access token. } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                [
                  empty_input,
                  input_too_long,
                  input_too_large,
                  unauthorized,
                  internal_error,
                  llm_unavailable,
                  billing_unavailable,
                  install_id_required,
                  invalid_purchase_body,
                  invalid_purchase,
                  store_unreachable,
                  license_not_ready,
                ]
            message: { type: string }
...